+++ NEWS.md
@@ -1,5 +1,8 @@
 # recipes (development version)
 
+* A new set of basis functions were added: `step_spline_b()`, `step_spline_convex()`,  `step_spline_monotone()`, `step_spline_natural()`, `step_spline_nonnegative()`, and 
+`step_poly_bernstein()`.
+
 # recipes 1.0.1
 
 * Fixed bug where `step_holiday()` didn't work if it isn't have any missing values. (#1019)